Start of refactoring of  package
