+++ source/core/src/main/com/csse3200/game/areas/OceanGameAreaConfigs/OceanSpawnConfig.java
@@ -1,4 +1,4 @@
 package com.csse3200.game.areas.OceanGameAreaConfigs;
 
-public class EntitySpawnConfig2 {
+public class OceanSpawnConfig {
 }
