+++ source/core/src/main/com/csse3200/game/areas/terrain/TerrainComponent.java
@@ -1,29 +1,18 @@
 package com.csse3200.game.areas.terrain;
 import java.util.*;
 
-import com.csse3200.game.areas.terrain.tiles.ForestTileConfig;
-import com.csse3200.game.areas.terrain.tiles.ForestTileConfig.TileConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.badlogic.gdx.graphics.OrthographicCamera;
-import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.maps.tiled.TiledMap;
 import com.badlogic.gdx.maps.tiled.TiledMapRenderer;
 import com.badlogic.gdx.maps.tiled.TiledMapTileLayer;
 import com.badlogic.gdx.math.GridPoint2;
 import com.badlogic.gdx.math.Vector2;
 import com.csse3200.game.areas.MapHandler.MapType;
-import com.csse3200.game.areas.FogGameAreaConfigs.FogMapTiles;
-import com.csse3200.game.areas.FogGameAreaConfigs.FogTileConfig;
-import com.csse3200.game.areas.OceanGameAreaConfigs.OceanMapTiles;
-import com.csse3200.game.areas.OceanGameAreaConfigs.OceanTileConfig;
-import com.csse3200.game.files.FileLoader;
 import com.csse3200.game.rendering.RenderComponent;
-import com.csse3200.game.services.ResourceService;
-import com.csse3200.game.services.ServiceLocator;
 
 /**
  * Render a tiled terrain for a given tiled map and orientation. A terrain is a
