+++ source/core/src/main/com/csse3200/game/areas/terrain/tiles/Tile.java
@@ -23,7 +23,7 @@ public class Tile {
     private BitSet down = new BitSet();
     private BitSet left = new BitSet();
 
-    public boolean collapsed = false;
+    private boolean isCollapsed = false;
 
     public Tile(String name, TextureRegion texture, List<String> edgeTiles, String centre) {
         this.name = name;
@@ -139,4 +139,7 @@ public class Tile {
     public void setPossibleLeft(BitSet left) {
             this.left = left;
         }
+
+    public boolean isCollapsed() {return isCollapsed;}
+    public void setIsCollapsed(boolean isCollapsed) {this.isCollapsed = isCollapsed;}
 }
